refactor: move date utils
- Move `/jira/utils.py` to `/utils/date.py`: It is universal
  - Remove `get_mixin_method`: This function is unused and may cause
    terrible redefining of mixin functions
  - Remove `escape_jql_string`: This function is unused
- Add export in `/utils/__init__.py`
- Remove `CommentsMixin._parse_date`: It is not suppposed to handle
  date format according to Single Responsibility Principle
