+++ src/mcp_atlassian/jira/worklog.py
@@ -5,8 +5,8 @@ import re
 from typing import Any
 
 from ..models import JiraWorklog
+from ..utils import parse_date_ymd
 from .client import JiraClient
-from .utils import parse_date_ymd
 
 logger = logging.getLogger("mcp-jira")
