+++ src/mcp_atlassian/utils/__init__.py
@@ -3,15 +3,11 @@ Utility functions for the MCP Atlassian integration.
 This package provides various utility functions used throughout the codebase.
 """
 
-# Re-export from ssl module
-# Re-export from io module
+# Re-export from modules
+from .date import parse_date, parse_date_human_readable, parse_date_ymd
 from .io import is_read_only_mode
-
-# Export new logging utilities
 from .logging import setup_logging
 from .ssl import SSLIgnoreAdapter, configure_ssl_verification
-
-# Re-export from urls module
 from .urls import is_atlassian_cloud_url
 
 # Export all utility functions for backward compatibility
@@ -20,5 +16,8 @@ __all__ = [
     "configure_ssl_verification",
     "is_atlassian_cloud_url",
     "is_read_only_mode",
+    "parse_date",
+    "parse_date_human_readable",
+    "parse_date_ymd",
     "setup_logging",
 ]
